useUser is only used by NavBar, move it to folder
